+++ frontend/src/App.css
@@ -160,3 +160,11 @@ hr {
     width: 18px;
     height: 18px;
 }
+
+.mobileScrollSpy {
+    padding: 1rem 0;
+    z-index: 99;
+    background-color: rgb(var(--bs-dark-rgb));
+    bottom: 0;
+    left: 0;
+}
