--- frontend/src/Pages/TakeAction/ScrollSpy/DesktopScrollSpy/ProgressNavLink.module.css
@@ -1,24 +1,0 @@
-.inactiveLink {
-    pointer-events: none;
-    cursor: default;
-}
-
-.numbers,
-.bullets,
-.checkmark {
-    font-weight: 600;
-    top: 50%;
-    transform: translate(0, -50%);
-    font-size: 0.9rem;
-    display: flex;
-    margin: auto 0;
-    position: absolute;
-}
-
-.checkmark {
-    color: #28a745;
-}
-
-.bullets {
-    font-size: 0.7rem;
-}
